# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/spaces/{space_id}/rooms/": {
      "get": {
        "summary": "Get Rooms",
        "operationId": "get_rooms_spaces__space_id__rooms__get",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  },
                  "title": "Response Get Rooms Spaces  Space Id  Rooms  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Rooms",
        "operationId": "delete_rooms_spaces__space_id__rooms__delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomsDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDeleteResDto"
                  },
                  "title": "Response Delete Rooms Spaces  Space Id  Rooms  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{space_id}/rooms/{room_id}": {
      "get": {
        "summary": "Get Room Information",
        "operationId": "get_room_information_spaces__space_id__rooms__room_id__get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomInfoDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/spaces/{space_id}/users/": {
      "get": {
        "summary": "Get Users",
        "operationId": "get_users_spaces__space_id__users__get",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  },
                  "title": "Response Get Users Spaces  Space Id  Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Users",
        "operationId": "create_users_spaces__space_id__users__post",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersImportRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportedUserResultDto"
                  },
                  "title": "Response Create Users Spaces  Space Id  Users  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Users",
        "operationId": "delete_users_spaces__space_id__users__delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDeleteResDto"
                  },
                  "title": "Response Delete Users Spaces  Space Id  Users  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{space_id}/users/{user_id}": {
      "get": {
        "summary": "Get User Information",
        "operationId": "get_user_information_spaces__space_id__users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/spaces/{space_id}/user_room/": {
      "post": {
        "summary": "Add Users To Rooms",
        "operationId": "add_users_to_rooms_spaces__space_id__user_room__post",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersAndRoomsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersAndRoomResDto"
                  },
                  "title": "Response Add Users To Rooms Spaces  Space Id  User Room  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{space_id}/user_room/group": {
      "delete": {
        "summary": "Remove Users From Room",
        "operationId": "remove_users_from_room_spaces__space_id__user_room_group_delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersAndRoomsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersAndRoomResDto"
                  },
                  "title": "Response Remove Users From Room Spaces  Space Id  User Room Group Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{space_id}/user_room/team": {
      "delete": {
        "summary": "Remove Users From Team",
        "operationId": "remove_users_from_team_spaces__space_id__user_room_team_delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersAndTeamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersAndTeamResDto"
                  },
                  "title": "Response Remove Users From Team Spaces  Space Id  User Room Team Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportedUserResultDto": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/UserCreateDto"
          },
          "created_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "email_sent": {
            "type": "boolean",
            "title": "Email Sent",
            "default": false
          },
          "email_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Error"
          }
        },
        "type": "object",
        "required": [
          "request",
          "password"
        ],
        "title": "ImportedUserResultDto"
      },
...
      "RoomDeleteResDto": {
        "properties": {
          "room": {
            "type": "string",
            "title": "Room"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "room"
        ],
        "title": "RoomDeleteResDto"
      },
      "RoomDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "broadcast": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broadcast"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "t": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "T"
          },
          "msgs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msgs"
          },
          "usersCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userscount"
          },
          "u": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShortUserDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "ro": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ro"
          },
          "default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "announcement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Announcement"
          }
        },
        "type": "object",
        "required": [
          "_id"
        ],
        "title": "RoomDto"
      },
      "RoomInfoDto": {
        "properties": {
          "team": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShortTeamDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "members": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RoomUserDto"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members"
          }
        },
        "type": "object",
        "title": "RoomInfoDto"
      },
      "RoomUserDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "username",
          "name",
          "status"
        ],
        "title": "RoomUserDto"
      },
      "RoomsDeleteDto": {
        "properties": {
          "rooms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rooms"
          }
        },
        "type": "object",
        "required": [
          "rooms"
        ],
        "title": "RoomsDeleteDto"
      },
...
      "ShortTeamDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "roomId": {
            "type": "string",
            "title": "Roomid"
          },
          "type": {
            "type": "integer",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "name",
          "roomId",
          "type"
        ],
        "title": "ShortTeamDto"
      },
      "ShortUserDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "username"
        ],
        "title": "ShortUserDto"
      },
...
      "UserCreateDto": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "name"
        ],
        "title": "UserCreateDto"
      },
      "UserDeleteResDto": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User"
          },
          "force_delete": {
            "type": "boolean",
            "title": "Force Delete"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "user",
          "force_delete"
        ],
        "title": "UserDeleteResDto"
      },
      "UserDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "avatarETag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avataretag"
          },
          "nameInsensitive": {
            "type": "string",
            "title": "Nameinsensitive"
          },
          "emails": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/UserEmailDto"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emails"
          },
          "lastLogin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastlogin"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "name",
          "username",
          "status",
          "active",
          "type",
          "roles",
          "nameInsensitive"
        ],
        "title": "UserDto"
      },
      "UserEmailDto": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "address",
          "verified"
        ],
        "title": "UserEmailDto"
      },
      "UserInfoDto": {
        "properties": {
          "teams": {
            "items": {
              "$ref": "#/components/schemas/ShortTeamDto"
            },
            "type": "array",
            "title": "Teams"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/UserInfoRoomDto"
            },
            "type": "array",
            "title": "Rooms"
          }
        },
        "type": "object",
        "required": [
          "teams",
          "rooms"
        ],
        "title": "UserInfoDto"
      },
      "UserInfoRoomDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rid": {
            "type": "string",
            "title": "Rid"
          },
          "t": {
            "type": "string",
            "title": "T"
          },
          "roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "name",
          "rid",
          "t"
        ],
        "title": "UserInfoRoomDto"
      },
      "UsersAndRoomResDto": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "user_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "User List"
          },
          "room": {
            "type": "string",
            "title": "Room"
          }
        },
        "type": "object",
        "required": [
          "user_list",
          "room"
        ],
        "title": "UsersAndRoomResDto"
      },
      "UsersAndRoomsDto": {
        "properties": {
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          },
          "rooms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rooms"
          }
        },
        "type": "object",
        "required": [
          "users",
          "rooms"
        ],
        "title": "UsersAndRoomsDto"
      },
      "UsersAndTeamResDto": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "user_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "User List"
          },
          "team": {
            "type": "string",
            "title": "Team"
          }
        },
        "type": "object",
        "required": [
          "user_list",
          "team"
        ],
        "title": "UsersAndTeamResDto"
      },
      "UsersAndTeamsDto": {
        "properties": {
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/DeleteTeamDto"
            },
            "type": "array",
            "title": "Teams"
          },
          "ban_in_rooms": {
            "type": "boolean",
            "title": "Ban In Rooms",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "users",
          "teams"
        ],
        "title": "UsersAndTeamsDto"
      },
      "UsersDeleteDto": {
        "properties": {
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          },
          "force_delete": {
            "type": "boolean",
            "title": "Force Delete",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "users"
        ],
        "title": "UsersDeleteDto"
      },
      "UsersImportRequestDto": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserCreateDto"
            },
            "type": "array",
            "title": "Users"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          },
          "requirePasswordChange": {
            "type": "boolean",
            "title": "Requirepasswordchange",
            "default": false
          },
          "joinDefaultChannels": {
            "type": "boolean",
            "title": "Joindefaultchannels",
            "default": true
          },
          "sendEmail": {
            "type": "boolean",
            "title": "Sendemail",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "users"
        ],
        "title": "UsersImportRequestDto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}